+++ R/vignettes.R
@@ -40,6 +40,7 @@ build_vignettes <- function(
     return(invisible())
   }
 
+  check_installed("remotes")
   deps <- remotes::dev_package_deps(pkg$path, dependencies)
   update(deps, upgrade = upgrade)
 
@@ -122,7 +123,7 @@ clean_vignettes <- function(pkg = ".") {
     file_delete(to_remove)
   }
 
-  lapply(c(doc_path, meta_path), dir_delete_if_empty)
+  walk(c(doc_path, meta_path), dir_delete_if_empty)
 
   invisible(TRUE)
 }